+++ CHANGELOG.md
@@ -8,12 +8,19 @@
 
 ### Improvements
 
+## v0.2.0 (2019.01.02)
+
+#### Features
+
+- Add ToS page
+- Add Button, Spinner components
+
 ## v0.1.2 (2019.01.02)
 
 #### Features
 
 - Able launch remote debug in vscode for main and renderer
-- Able to run `yarn cucumber` for integration test 
+- Able to run `yarn cucumber` for integration test
 - Added preload.js script for security
 - Able to run cennnznet-node as part of the app